OCR-to-CSV-Automation/requirements.txt
pandas
streamlit
pytesseract
pdf2image
Pillow
scikit-learn
